Move updateAndFireQueueStatus method to DownloadCacheImpl.java
